refactor (be): package filters together
